+++ CHANGELOG.md
@@ -17,6 +17,8 @@ $ cd autohooks && git log
 ## [Unreleased]
 ### Added
 ### Changed
+* Using `Path.cwd()` instead of `os.environ['PWD']`, to make autohooks more reliable for windows. [#165](https://github.com/greenbone/autohooks/pull/165)
+
 ### Deprecated
 ### Removed
 ### Fixed